feat: Add Support for Tasklist (#44)
* feat: added task service and context layer

* added swipeable list dependency

* add TaskProvider in App

* modify api-service to handle config and remove payload

* modified apptheme

* modify taskcontext to handle loading state on error

* constants added

* modify task-service

* added task section in dashboard

* refactor: handle error message

* optimized code

* added validation message in locale instead of constants

* modified services to use bearer header in base service

* renamed explicitly named functions to general function names

* removed explicit type conversion

* added empty line at end

* renamed function name

* added task operation messages in locale

+++ src/contexts/account-context.tsx
@@ -8,7 +8,7 @@ import React, {
 } from 'react';
 
 import { AccountService } from '../services';
-import { AccessToken, Account, Nullable } from '../types';
+import { Account, Nullable } from '../types';
 
 import { useAuthContext } from './auth-context';
 
@@ -38,13 +38,11 @@ export const AccountContextProvider: React.FC<PropsWithChildren> = ({ children }
   const [isUpdateAccountLoading, setIsUpdateAccountLoading] = useState(false);
   const [isDeleteAccountLoading, setIsDeleteAccountLoading] = useState(false);
 
-  const { getAccessTokenFromStorage } = useAuthContext();
+  const { getAccessToken } = useAuthContext();
 
   const getAccountDetails = async () => {
     setIsAccountLoading(true);
-    const { data, error } = await accountService.getAccount(
-      getAccessTokenFromStorage() as AccessToken,
-    );
+    const { data, error } = await accountService.getAccount(getAccessToken());
     if (data) {
       setAccountDetails(new Account({ ...data }));
       if (!data.firstName) {
@@ -64,7 +62,7 @@ export const AccountContextProvider: React.FC<PropsWithChildren> = ({ children }
     const { data, error } = await accountService.updateAccount(
       firstName,
       lastName,
-      getAccessTokenFromStorage() as AccessToken,
+      getAccessToken(),
     );
     if (data) {
       setAccountDetails(new Account({ ...data }));
@@ -77,9 +75,7 @@ export const AccountContextProvider: React.FC<PropsWithChildren> = ({ children }
 
   const deleteAccount = async () => {
     setIsDeleteAccountLoading(true);
-    const { error } = await accountService.deleteAccount(
-      getAccessTokenFromStorage() as AccessToken,
-    );
+    const { error } = await accountService.deleteAccount(getAccessToken());
     if (error) {
       setIsDeleteAccountLoading(false);
       throw error;
